Moved the lookup generation script to the top level but kept utility package in nwis_code_lookups. Codes are now saved to data directory
