+++ .gitignore
@@ -1,6 +1,8 @@
 #Pycharm/IDE specific files
 .idea/
 
+data/
+
 # Byte-compiled / optimized / DLL files
 __pycache__/
 *.py[cod]
