--- bin/nwis_code_lookups/requirements.txt
@@ -1,5 +1,0 @@
-certifi==2017.11.5
-chardet==3.0.4
-idna==2.6
-requests==2.18.4
-urllib3==1.22
